add curator

+++ client/src/main/java/com/example/client/config/ZkReenterableLock.java
@@ -1,24 +1,23 @@
 package com.example.client.config;
 
 
-import org.apache.zookeeper.*;
+import org.apache.zookeeper.CreateMode;
+import org.apache.zookeeper.KeeperException;
+import org.apache.zookeeper.ZooDefs;
+import org.apache.zookeeper.ZooKeeper;
 import org.apache.zookeeper.data.Stat;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.DependsOn;
-import org.springframework.core.annotation.Order;
-import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.locks.Condition;
 import java.util.concurrent.locks.Lock;
 
-public class ZkLock implements Lock {
+public class ZkReenterableLock implements Lock {
 
   private String root = "/locks";
   private String lockName ;
@@ -28,10 +27,11 @@ public class ZkLock implements Lock {
   private CountDownLatch latch ;
   private ZooKeeper zkClient;
 //  锁可重入设计
-  private volatile Thread curthread;
-  private AtomicInteger lockcount = new AtomicInteger(0);
+  private static volatile Thread curthread;
+  private static AtomicInteger lockcount = new AtomicInteger(0);
 
-  public ZkLock(ZooKeeper zkClient,String lockName) throws KeeperException, InterruptedException {
+
+  public ZkReenterableLock(ZooKeeper zkClient, String lockName) throws KeeperException, InterruptedException {
     this.zkClient = zkClient;
     this.lockName = lockName;
     Stat stat = zkClient.exists(root, false);
@@ -40,13 +40,11 @@ public class ZkLock implements Lock {
     }
   }
 
-
   @Override
   public void lock() {
     if(tryLock()){
       System.out.println("当前线程"+Thread.currentThread()+"node:"+myNode);
-      this.curthread = Thread.currentThread();
-      lockcount.incrementAndGet();
+
     }else{
       try {
         waitForLock(preNode,10);
@@ -69,24 +67,32 @@ public class ZkLock implements Lock {
   @Override
   public boolean tryLock() {
     try {
-      if(myNode == null){
-        myNode = zkClient.create(root+"/"+lockName+splitstr, new byte[0], ZooDefs.Ids.OPEN_ACL_UNSAFE, CreateMode.EPHEMERAL_SEQUENTIAL);
-      }
-      List<String> children = zkClient.getChildren(root, false);
-      List<String> lockNamechildrens = new ArrayList();
-      children.forEach( c -> {
-        if(c.split(splitstr)[0].equals(lockName)){
-          lockNamechildrens.add(c);
+      if(lockcount.get() == 0){
+        if(myNode == null){
+          myNode = zkClient.create(root+"/"+lockName+splitstr, new byte[0], ZooDefs.Ids.OPEN_ACL_UNSAFE, CreateMode.EPHEMERAL_SEQUENTIAL);
         }
-      });
-      Collections.sort(lockNamechildrens);
-      if(!lockNamechildrens.isEmpty() && myNode.equals(root+"/"+lockNamechildrens.get(0))){
+        List<String> children = zkClient.getChildren(root, false);
+        List<String> lockNamechildrens = new ArrayList();
+        children.forEach( c -> {
+          if(c.split(splitstr)[0].equals(lockName)){
+            lockNamechildrens.add(c);
+          }
+        });
+        Collections.sort(lockNamechildrens);
+        if(!lockNamechildrens.isEmpty() && myNode.equals(root+"/"+lockNamechildrens.get(0))){
+          this.curthread = Thread.currentThread();
+          lockcount.incrementAndGet();
+          return true;
+        }
+        //当前节点没抢到，把前一个节点找出来
+        String subMyNode = myNode.substring(myNode.lastIndexOf("/")+1);
+        int mynodeindex = Collections.binarySearch(lockNamechildrens, subMyNode);
+        preNode = lockNamechildrens.get(mynodeindex - 1);
+      }else if(this.curthread == Thread.currentThread()){
+        lockcount.incrementAndGet();
         return true;
       }
-      //当前节点没抢到，把前一个节点找出来
-      String subMyNode = myNode.substring(myNode.lastIndexOf("/")+1);
-      int mynodeindex = Collections.binarySearch(lockNamechildrens, subMyNode);
-      preNode = lockNamechildrens.get(mynodeindex - 1);
+
     } catch (KeeperException e) {
       e.printStackTrace();
     } catch (InterruptedException e) {
@@ -116,8 +122,12 @@ public class ZkLock implements Lock {
   @Override
   public void unlock() {
     try {
-      zkClient.delete(myNode,-1);
-//      zkClient.close();
+      if(Thread.currentThread() != curthread)return;
+      int i = lockcount.decrementAndGet();
+      if(i == 0){
+        curthread = null;
+        zkClient.delete(myNode,-1);
+      }
       System.out.println(Thread.currentThread()+"释放锁");
     } catch (Exception e) {
       e.printStackTrace();
